+++ data/account.mock.ts
@@ -1,0 +1,5 @@
+const ACCOUNT_TYPE = "personal"
+
+export {
+	ACCOUNT_TYPE
+}
