+++ data/carrier.mock.ts
@@ -1,0 +1,106 @@
+const MOCKUP_TRUCK = [
+	{
+		truck_id: 0,
+		license_number: "90-6179",
+		wheel: "หัวลาก",
+		add_on: "2 เพลา",
+		status: 100
+	},
+	{
+		truck_id: 1,
+		license_number: "89-7280",
+		wheel: "6 ล้อ",
+		add_on: "-",
+		status: 200
+	},
+	{
+		truck_id: 2,
+		license_number: "กข-1111",
+		wheel: "4 ล้อ",
+		add_on: "ตู้ทึบ",
+		status: 300
+	},
+]
+
+const MOCKUP_DRIVER = [
+	{
+		id: "01",
+		driver_name: "อสมาภรณ์ ดีวรัตน์",
+		driver_license_type: "ท.2",
+	},
+	{
+		id: "02",
+		driver_name: "ปูรณ์ โชตธีรวสุ",
+		driver_license_type: "ท.4",
+	},
+	{
+		id: "03",
+		driver_name: "ชัญญา พุกกะรัตน​์",
+		driver_license_type: "ท.2",
+	},
+	{
+		id: "04",
+		driver_name: "ชิษณุพงศ์ วรวิจิตรชัยกุล",
+		driver_license_type: "ท.3",
+	},
+	{
+		id: "05",
+		driver_name: "จิรัชฌา วิทยาชีวะ",
+		driver_license_type: "ท.3",
+	},
+	{
+		id: "06",
+		driver_name: "นรเศรษฐ์ โพธิ์ทอง",
+		driver_license_type: "ท.3",
+	},
+	{
+		id: "07",
+		driver_name: "ศิวกร ไทยแท้",
+		driver_license_type: "ท.4",
+	},
+	{
+		id: "08",
+		driver_name: "ฐิติวัชร์ นาพรหม",
+		driver_license_type: "ท.2",
+	},
+	{
+		id: "09",
+		driver_name: "ณัฐพัชร์ วัฒนชาญสิทธิ์",
+		driver_license_type: "ท.1",
+	},
+	{
+		id: "10",
+		driver_name: "ชัชวิน แท่งทอง",
+		driver_license_type: "ท.3",
+	},
+	{
+		id: "11",
+		driver_name: "จารวี ทองเจริญ",
+		driver_license_type: "ท.1",
+	},
+	{
+		id: "12",
+		driver_name: "กฤตพล จันทร์ศุภกุล",
+		driver_license_type: "ท.1",
+	},
+	{
+		id: "13",
+		driver_name: "ธีรธัชช์ นักร้อง",
+		driver_license_type: "ท.1",
+	},
+	{
+		id: "14",
+		driver_name: "ปัณฑา ปราโมทย์กุล",
+		driver_license_type: "ท.3",
+	},
+	{
+		id: "15",
+		driver_name: "พรรษชล เผ่าทิตธรรม",
+		driver_license_type: "ท.4",
+	}
+]
+
+export {
+	MOCKUP_TRUCK,
+	MOCKUP_DRIVER
+}
