+++ data/carrier.ts
@@ -1,0 +1,73 @@
+const DRIVER_LICENSE_TYPE = ["ประเภท 1", "ประเภท 2 (บ.2 หรือ ท.2)", "ประเภท 3 (บ.3 หรือ ท.3)", "ประเภท 4"]
+
+const TRUCK_STATUS_LIST = {
+	0: "ทุกสถานะ",
+	100: "จอดว่าง",
+	200: "กำลังขนส่ง",
+	300: "ซ่อมบำรุง"
+}
+
+const TRUCK_TYPE = [
+	{
+		type: "รถ 4 ล้อ",
+		option: [
+			"ตู้ทึบ",
+			"เปิดข้าง",
+			"คอก",
+			"คอกสูง"
+		]
+	},
+	{
+		type: "รถ 6 ล้อ",
+		option: [
+			"ตู้ทึบ",
+			"เปิดข้าง",
+			"คอก",
+			"พื้นเรียบ",
+			"ตู้เย็น"
+		]
+	},
+	{
+		type: "รถ 10 ล้อ",
+		option: [
+			"ตู้คอนเทนเนอร์",
+			"เปิดข้าง",
+			"คอก",
+			"พื้นเรียบ",
+			"ตู้เย็น"
+		]
+	},
+	{
+		type: "หัวลาก 10 ล้อ",
+		option: [
+			"ตู้คอนเทนเนอร์",
+			"พื้นเรียบ",
+			"ตู้เย็น",
+			"ก้างปลา"
+		],
+		chassis: [
+			"2 เพลา",
+			"3 เพลา"
+		]
+	},
+	{
+		type: "หัวลาก 18 ล้อ",
+		option: [
+			"ตู้คอนเทนเนอร์",
+			"พื้นเรียบ",
+			"ตู้เย็น",
+			"ก้างปลา"
+		],
+		chassis: [
+			"2 เพลา",
+			"3 เพลา"
+		]
+	},
+]
+
+
+export {
+	DRIVER_LICENSE_TYPE,
+	TRUCK_TYPE,
+	TRUCK_STATUS_LIST
+}
